pyproject.toml
[project]
name = "deep-learning-project"
version = "0.1.0"
description = "Deep Learning project for the Sapienza's DL course"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/audio_sr"]

[tool.hatch.build]
sources = ["src"]
